virtual_tryon_project/requirements.txt
# Virtual Try-On System Requirements
# Python 3.10+

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Image Processing
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0
albumentations>=1.3.0

# Numerical Computing
numpy>=1.24.0,<2.0.0  # Force NumPy 1.x for compatibility with matplotlib/opencv
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Tracking
tqdm>=4.65.0

# Web Interface
streamlit>=1.25.0
gradio>=3.40.0

# Metrics
scikit-learn>=1.3.0
lpips>=0.1.4

# File I/O
python-dotenv>=1.0.0
pyyaml>=6.0

# Data Augmentation
imgaug>=0.4.0

# Optional: For better performance
# ninja>=1.11.0
